server/requirements.txt
fastapi[standard]
sqlmodel
argon2-cffi